Доработка проекта
Исправлены некоторые ошибки, улучшена обработка сигналов, убраны варнинги.

+++ FIRMWARE/SH1106/OLED.h
@@ -14,7 +14,7 @@ void PutChar_mid(unsigned char x,unsigned char y,unsigned char dig,unsigned char
 void PutStr_mid(unsigned char x, unsigned char y, const char txt[],unsigned char Adress);
 void PutNumber_mid(unsigned char x, unsigned char y, int val,unsigned char Adr);
 void OLED_temper_big(unsigned char x, unsigned char y,int val, unsigned char Adr);
-void OLED_4_big(unsigned char x, unsigned char y,unsigned int val, unsigned char Adr, unsigned char c);
+void OLED_4_big(unsigned char x, unsigned char y,unsigned long int val, unsigned char Adr, unsigned char c);
 void OLED_mid(unsigned char x, unsigned char y, int val,unsigned char Adr, unsigned char c);
 void OLED_dot_mid(unsigned char x, unsigned char y, unsigned char Adr);
 void OLED_doubledot_mid(unsigned char x, unsigned char y, unsigned char Adr);
